+++ Releasenotes.md
@@ -1,7 +1,15 @@
-# version 4.1.0
 
+# Version 4.2.0
+- bugfix: added functionality to retreive active network adapter
+- added logging of database credentials to /home/pi/db_setup_enecsys.log
+- code improvement: webinstaller
+- renamed installer files
+- documentation updated
+
+
+# version 4.1.0
 # 2019-11-25
-* Start with rebuild for Raspian Buster
+* Rebuilded for Raspbian Buster
 
 # Fixes:
 * Update installer scripts for Raspbian Buster